Truncate e2e tests, minor refactoring

+++ destination/main_e2e_test.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"context"
 	"encoding/csv"
 	"encoding/json"
 	"errors"
@@ -15,7 +16,10 @@ import (
 
 	"fivetran.com/fivetran_sdk/destination/cmd"
 	"fivetran.com/fivetran_sdk/destination/common/flags"
+	"fivetran.com/fivetran_sdk/destination/db"
 	"fivetran.com/fivetran_sdk/destination/db/config"
+	"fivetran.com/fivetran_sdk/destination/service"
+	pb "fivetran.com/fivetran_sdk/proto"
 	"github.com/stretchr/testify/require"
 )
 
@@ -29,7 +33,7 @@ func TestAllDataTypes(t *testing.T) {
 	fileName := "input_all_data_types.json"
 	tableName := "all_data_types"
 	startServer(t)
-	runSDKTestCommand(t, fileName)
+	runSDKTestCommand(t, fileName, true)
 	assertTableRowsWithFivetranId(t, tableName, [][]string{
 		{"true", "42", "144", "100500", "100.5", "200.5", "42.42",
 			"2024-05-07", "2024-04-05 15:33:14", "2024-02-03 12:44:22.123456789",
@@ -61,7 +65,7 @@ func TestMutateAfterAlter(t *testing.T) {
 	fileName := "input_mutate_after_alter.json"
 	tableName := "mutate_after_alter"
 	startServer(t)
-	runSDKTestCommand(t, fileName)
+	runSDKTestCommand(t, fileName, true)
 	assertTableRowsWithPK(t, tableName, [][]string{
 		{"1", "200", "asd", "zxc", "\\N", "\\N", "false"},
 		{"2", "50", "\\N", "\\N", "<c>99</c>", "DD", "false"},
@@ -82,7 +86,7 @@ func TestUpdateAndDelete(t *testing.T) {
 	fileName := "input_update_and_delete.json"
 	tableName := "update_and_delete"
 	startServer(t)
-	runSDKTestCommand(t, fileName)
+	runSDKTestCommand(t, fileName, true)
 	assertTableRowsWithPK(t, tableName, [][]string{
 		{"1", "1111", "false"},
 		{"2", "two", "false"},
@@ -100,7 +104,7 @@ func TestNonExistentRecordUpdatesAndDeletes(t *testing.T) {
 	fileName := "input_non_existent_updates.json"
 	tableName := "non_existent_updates"
 	startServer(t)
-	runSDKTestCommand(t, fileName)
+	runSDKTestCommand(t, fileName, true)
 	assertTableRowsWithPK(t, tableName, [][]string{
 		{"1", "\\N", "false"}})
 	assertTableColumns(t, tableName, [][]string{
@@ -110,12 +114,18 @@ func TestNonExistentRecordUpdatesAndDeletes(t *testing.T) {
 		{"_fivetran_deleted", "Bool", ""}})
 }
 
+// Currently, SDK tester will always send the truncate operation last, no matter its position in the input file,
+// and it is also a "soft" truncate only. See TestTruncateBefore for workaround to test "truncate before" + "hard".
+// See TestTruncateExistingRecordsThenSync that verifies if we can do a full sync after truncating the table.
 func TestTruncate(t *testing.T) {
 	fileName := "input_truncate.json"
 	tableName := "table_to_truncate"
 	startServer(t)
-	runSDKTestCommand(t, fileName)
-	assertTableRowsWithPK(t, tableName, nil)
+	runSDKTestCommand(t, fileName, true)
+	assertTableRowsWithPK(t, tableName, [][]string{
+		{"1", "foo", "true"},
+		{"2", "bar", "true"},
+		{"3", "qaz", "true"}})
 	assertTableColumns(t, tableName, [][]string{
 		{"id", "Int32", ""},
 		{"name", "Nullable(String)", ""},
@@ -123,19 +133,137 @@ func TestTruncate(t *testing.T) {
 		{"_fivetran_deleted", "Bool", ""}})
 }
 
+func TestTruncateBefore(t *testing.T) {
+	// FIXME:
+	//  should be possible to test it "properly" with future SDK releases, currently it's a workaround,
+	//  as truncate operation is always last and all the input.json rows are merged into one "replace" CSV file.
+	//  With a fixed version, direct conn calls should be replaced with SDK tester calls instead.
+	fileName := "input_truncate_create_table.json"
+	tableName := "table_to_truncate"
+
+	startServer(t)
+	conf := readConfigMap(t)
+	conn, err := db.GetClickHouseConnection(conf)
+	require.NoError(t, err)
+	defer conn.Close()
+
+	// create a table via SDK first, then store a record with _fivetran_synced in the future;
+	// verify that this record is not "soft deleted" after truncate
+	upsertRecordsWithSDK := func() {
+		runSDKTestCommand(t, fileName, true)
+		assertTableColumns(t, tableName, [][]string{
+			{"id", "Int32", ""},
+			{"name", "Nullable(String)", ""},
+			{"_fivetran_synced", "DateTime64(9, 'UTC')", ""},
+			{"_fivetran_deleted", "Bool", ""}})
+
+		syncTime := time.Now().Add(1 * time.Minute).UTC()
+		runQuery(t, fmt.Sprintf("INSERT INTO tester.%s VALUES (1, 'foo', %d, false)",
+			tableName, syncTime.UnixNano()))
+	}
+
+	/// soft (ALTER TABLE UPDATE)
+	upsertRecordsWithSDK()
+	err = conn.TruncateTable(context.Background(), schemaName, tableName, syncedColumn, time.Now(), &softDeletedColumn)
+	require.NoError(t, err)
+	assertTableRowsWithPK(t, tableName, [][]string{
+		{"1", "foo", "false"},
+		{"2", "bar", "true"},
+		{"3", "qaz", "true"}})
+
+	/// hard (ALTER TABLE DELETE)
+	upsertRecordsWithSDK()
+	err = conn.TruncateTable(context.Background(), schemaName, tableName, syncedColumn, time.Now(), nil)
+	require.NoError(t, err)
+	assertTableRowsWithPK(t, tableName, [][]string{
+		{"1", "foo", "false"},
+	})
+}
+
+// FIXME: should be possible to fully simulate using just SDK tester.
+func TestTruncateExistingRecordsThenSync(t *testing.T) {
+	fileName := "input_truncate_then_sync.json"
+	tableName := "truncate_then_sync"
+
+	startServer(t)
+	conf := readConfigMap(t)
+	conn, err := db.GetClickHouseConnection(conf)
+	require.NoError(t, err)
+	defer conn.Close()
+
+	runSDKTestCommand(t, fileName, true)
+	assertTableRowsWithPK(t, tableName, [][]string{
+		{"1", "name-truncated-1", "desc-truncated-1", "true"},
+		{"2", "name-truncated-2", "desc-truncated-2", "true"}})
+	assertTableColumns(t, tableName, [][]string{
+		{"id", "Int32", ""},
+		{"name", "Nullable(String)", ""},
+		{"desc", "Nullable(String)", ""},
+		{"_fivetran_synced", "DateTime64(9, 'UTC')", ""},
+		{"_fivetran_deleted", "Bool", ""}})
+
+	ctx := context.Background()
+	table := &pb.Table{
+		Name: tableName,
+		Columns: []*pb.Column{
+			{Name: "id", Type: pb.DataType_INT, PrimaryKey: true},
+			{Name: "name", Type: pb.DataType_STRING, PrimaryKey: false},
+			{Name: "desc", Type: pb.DataType_STRING, PrimaryKey: false},
+			{Name: "_fivetran_synced", Type: pb.DataType_UTC_DATETIME, PrimaryKey: false},
+			{Name: "_fivetran_deleted", Type: pb.DataType_BOOLEAN, PrimaryKey: false},
+		}}
+
+	metadata, err := service.GetPrimaryKeysAndMetadataColumns(table)
+	require.NoError(t, err)
+
+	colTypes, err := conn.GetColumnTypes(ctx, schemaName, tableName)
+	require.NoError(t, err)
+
+	nullStr := "this-is-null"
+	unmodifiedStr := "do-not-modify"
+	syncTime := time.Now().Format("2006-01-02T15:04:05.000000000Z")
+	replaceCSV := [][]string{
+		{"1", "name-replaced-1", "desc-replaced-1", syncTime, "false"},
+		{"2", "name-replaced-2", "desc-replaced-2", syncTime, "false"},
+	}
+	updateCSV := [][]string{
+		{"1", unmodifiedStr, "desc-updated-1", syncTime, "false"},
+	}
+	deleteCSV := [][]string{
+		{"2", nullStr, nullStr, syncTime, "true"},
+	}
+
+	err = conn.ReplaceBatch(ctx, schemaName, table, replaceCSV, nullStr, 100)
+	require.NoError(t, err)
+	err = conn.UpdateBatch(ctx, schemaName, table, metadata.PrimaryKeys, colTypes, updateCSV, nullStr, unmodifiedStr, 100, 100, 5)
+	require.NoError(t, err)
+	err = conn.SoftDeleteBatch(ctx, schemaName, table, metadata.PrimaryKeys, colTypes, deleteCSV, metadata.FivetranSyncedIdx, metadata.FivetranDeletedIdx, 100, 100, 5)
+	require.NoError(t, err)
+
+	// FIXME: this is actually wrong.
+	//  While the final values are correct, we are missing previous "soft truncated" records.
+	//  Those were replaced with their newer versions via ReplacingMergeTree.
+	assertTableRowsWithPK(t, tableName, [][]string{
+		{"1", "name-replaced-1", "desc-updated-1", "false"},
+		{"2", "name-replaced-2", "desc-replaced-2", "true"}})
+}
+
 func TestTableNotFound(t *testing.T) {
 	fileName := "input6_table_not_found.json"
 	startServer(t)
-	runSDKTestCommand(t, fileName) // verify at least no SDK tester errors
+	runSDKTestCommand(t, fileName, true) // verify at least no SDK tester errors
 }
 
 func TestLargeInputFile(t *testing.T) {
-	fileName := "input_large_file"
+	tableName := "input_large_file"
 	startServer(t)
-	expectedCSV := generateAndWriteInputFile(t, fileName, 150_000)
-	runSDKTestCommand(t, fmt.Sprintf("%s.json", fileName))
-	assertTableRowsWithPK(t, fileName, expectedCSV)
-	assertTableColumns(t, fileName, [][]string{
+
+	expectedCSV := generateAndWriteInputFile(t, tableName, 150_000)
+	runSDKTestCommand(t, fmt.Sprintf("%s.json", tableName), true)
+
+	dbRecordsCSVStr := runQuery(t, fmt.Sprintf("SELECT * EXCEPT _fivetran_synced FROM tester.%s FINAL ORDER BY id FORMAT CSV", tableName))
+	assertDatabaseRecordsFailFast(t, expectedCSV, dbRecordsCSVStr)
+	assertTableColumns(t, tableName, [][]string{
 		{"id", "Int64", ""},
 		{"data", "Nullable(String)", ""},
 		{"created_at", "Nullable(DateTime)", ""},
@@ -143,7 +271,8 @@ func TestLargeInputFile(t *testing.T) {
 		{"_fivetran_deleted", "Bool", ""}})
 }
 
-func assertDatabaseRecords(t *testing.T, expectedRecords [][]string, dbRecordsCSVStr string) {
+// fail on the first mismatch, to prevent long console output
+func assertDatabaseRecordsFailFast(t *testing.T, expectedRecords [][]string, dbRecordsCSVStr string) {
 	csvReader := csv.NewReader(strings.NewReader(dbRecordsCSVStr))
 	for _, expected := range expectedRecords {
 		dbRecord, _ := csvReader.Read()
@@ -151,14 +280,20 @@ func assertDatabaseRecords(t *testing.T, expectedRecords [][]string, dbRecordsCS
 	}
 }
 
+func assertDatabaseRecords(t *testing.T, expectedRecords [][]string, dbRecordsCSVStr string) {
+	dbRecords, err := csv.NewReader(strings.NewReader(dbRecordsCSVStr)).ReadAll()
+	require.NoError(t, err)
+	require.Equal(t, expectedRecords, dbRecords)
+}
+
 func assertTableRowsWithPK(t *testing.T, tableName string, expectedOutput [][]string) {
-	query := fmt.Sprintf("SELECT * EXCEPT _fivetran_synced FROM tester.%s FINAL ORDER BY id FORMAT CSV", tableName)
+	query := fmt.Sprintf("SELECT * EXCEPT _fivetran_synced FROM tester.%s FINAL ORDER BY id FORMAT CSV SETTINGS select_sequential_consistency=1", tableName)
 	dbRecordsCSVStr := runQuery(t, query)
 	assertDatabaseRecords(t, expectedOutput, dbRecordsCSVStr)
 }
 
 func assertTableRowsWithFivetranId(t *testing.T, tableName string, expectedOutput [][]string) {
-	query := fmt.Sprintf("SELECT * EXCEPT _fivetran_synced FROM tester.%s FINAL ORDER BY _fivetran_id FORMAT CSV", tableName)
+	query := fmt.Sprintf("SELECT * EXCEPT _fivetran_synced FROM tester.%s FINAL ORDER BY _fivetran_id FORMAT CSV SETTINGS select_sequential_consistency=1", tableName)
 	dbRecordsCSVStr := runQuery(t, query)
 	assertDatabaseRecords(t, expectedOutput, dbRecordsCSVStr)
 }
@@ -254,6 +389,7 @@ func generateAndWriteInputFile(t *testing.T, tableName string, n uint) [][]strin
 const dialTimeout = 10 * time.Millisecond
 const maxDialRetries = 300
 
+var configMap atomic.Value
 var connConfig atomic.Value
 
 func startServer(t *testing.T) {
@@ -275,9 +411,9 @@ func getProjectRootDir(t *testing.T) string {
 	return result
 }
 
-func readConfig(t *testing.T) *config.Config {
-	if connConfig.Load() != nil {
-		return connConfig.Load().(*config.Config)
+func readConfigMap(t *testing.T) map[string]string {
+	if configMap.Load() != nil {
+		return configMap.Load().(map[string]string)
 	}
 	rootDir := getProjectRootDir(t)
 	configBytes, err := os.ReadFile(fmt.Sprintf("%s/sdk_tests/configuration.json", rootDir))
@@ -286,6 +422,15 @@ func readConfig(t *testing.T) *config.Config {
 	m := make(map[string]string)
 	err = json.Unmarshal(configBytes, &m)
 	require.NoError(t, err)
+	configMap.Store(m)
+	return m
+}
+
+func readConfig(t *testing.T) *config.Config {
+	if connConfig.Load() != nil {
+		return connConfig.Load().(*config.Config)
+	}
+	m := readConfigMap(t)
 	res, err := config.Parse(m)
 	require.NoError(t, err)
 	connConfig.Store(res)
@@ -340,9 +485,11 @@ func waitPortIsReady(t *testing.T, port uint) {
 	t.Fatalf("Port is not ready after %d retries", maxDialRetries)
 }
 
-func runSDKTestCommand(t *testing.T, inputFileName string) {
-	runQuery(t, "DROP DATABASE IF EXISTS tester")
-	runQuery(t, "CREATE DATABASE IF NOT EXISTS tester")
+func runSDKTestCommand(t *testing.T, inputFileName string, recreateDatabase bool) {
+	if recreateDatabase {
+		runQuery(t, "DROP DATABASE IF EXISTS tester")
+		runQuery(t, "CREATE DATABASE IF NOT EXISTS tester")
+	}
 	projectRootDir := getProjectRootDir(t)
 	command := exec.Command("make", "sdk-test")
 	command.Dir = projectRootDir
@@ -357,3 +504,7 @@ func runSDKTestCommand(t *testing.T, inputFileName string) {
 	out := string(byteOut)
 	require.Contains(t, out, "[Test connection and basic operations]: PASSED")
 }
+
+var schemaName = "tester"
+var softDeletedColumn = "_fivetran_deleted"
+var syncedColumn = "_fivetran_synced"
